refactor(1011): extract WANProbeConfigManager to src/refactors/wanprobe_config_manager.py (SC-015)
Ports the 473 LoC WANProbeConfigManager class from MistHelper.py (menu 166
WAN interface ICMP probe configurator) into its own module under
src/refactors/. Applies the late-binding _MistHelperProxy/_MH pattern to
resolve MistHelper-owned dependencies (ConfigUtils, InputUtils, CacheUtils,
GatewayExportUtils, OrgSiteExporter, FilePathUtils, DataExporter,
MIST_SITE_EXCLUDE_PREFIX, apisession) via importlib on each attribute
access, avoiding circular imports and preserving live rebindings.

Landing gates: py_compile OK, ruff clean, black applied, mypy --strict
clean, pydocstyle exit 0, compliance analyzer 100.0/A+ on the new
module. MistHelper.py compliance preserved at 93.0/A. Refactor catalog
regenerated: low-use bucket now 18 (was 19).

MistHelper.py: 22,933 -> 22,456 lines (-477).
